Remove duplicated annotations and put all in the ObjectBase.
